Change semantics of calls to execution context built-in methods from the body of multifile predicate clauses
Calls to the `sender/1`, `this/1`, and `self/1` execution context built-in methods from the body of multifile predicate clauses now always return the entity containing the clause. This semantic change is necessary to ensure that any direct or indirect call (in particular, sending a message that calls a meta-predicate) made from the clause body will use the expected execution context.